Added functionaltyto CUST INFO and MAILING ADDRESS to handle double-quotes in the strings.
